Adcionando condicionais às funcionalidades da API

+++ servidor.js
@@ -25,6 +25,7 @@ const sql = neon(process.env.DATABASE_URL);
 // Lista alimentos
 fastify.get('/alimentos', async (request, reply) => {
     const alimentos = await sql `SELECT * FROM alimentos`;
+    if (alimentos.length === 0) {return reply.status(404).send({erro: "Nenhum alimento encontrado"})}
     return reply.status(200).send(alimentos)
 })
 
@@ -36,6 +37,9 @@ fastify.get('/alimentos', async (request, reply) => {
 fastify.post("/alimentos", async (request, reply) => {
     const {nome, peso, ptn, carb, fat, kcal} = request.body 
 
+    if (!nome || !peso || !ptn || !carb || !fat || !kcal) {
+        return reply.status(400).send({ message: "Todos os campos são obrigatórios" });}
+    
     await sql `INSERT INTO alimentos (nome, peso, ptn, carb, fat, kcal) 
     VALUES (${nome}, ${peso}, ${ptn}, ${carb}, ${fat}, ${kcal})`;
 
@@ -52,8 +56,12 @@ fastify.put("/alimentos/:id", async (request, reply) => {
     const alimentoID = request.params.id
     const {nome, peso, ptn, carb, fat, kcal} = request.body 
 
+    if (!alimentoID) {return reply.status(404).send({erro: "Alimento não encontrado"})}
+    else if (!nome || !peso || !ptn || !carb || !fat || !kcal) {
+        return reply.status(400).send({ message: "Todos os campos são obrigatórios" });}
+
     await sql `UPDATE alimentos set nome = ${nome}, ${peso}, 
-    ${ptn}, ${carb}, ${fat} WHERE id = ${alimentoID}`
+    ${ptn}, ${carb}, ${fat} WHERE id = ${alimentoID}`;
 
     return reply.status(204).send({message: "Alimento atualizado com suceso"})
 })
@@ -64,6 +72,8 @@ fastify.put("/alimentos/:id", async (request, reply) => {
 
 fastify.delete("/alimentos/:id", async (request, reply) => {
     const alimentoID = request.params
+    
+    if (!alimentoID) {return reply.status(404).send({erro: "Alimento não encontrado"})}
 
     await sql `DELETE FROM alimentos WHERE id = ${alimentoID}`
